+++ EasyPost/Exceptions/API/UnknownHttpError.cs
@@ -3,16 +3,16 @@ using EasyPost.Models.API;
 
 namespace EasyPost.Exceptions.API
 {
-    public class UnexpectedHttpError : ApiError
+    public class UnknownHttpError : ApiError
     {
         /// <summary>
-        ///     Initializes a new instance of the <see cref="UnexpectedHttpError" /> class.
+        ///     Initializes a new instance of the <see cref="UnknownHttpError" /> class.
         /// </summary>
         /// <param name="errorMessage">Error message string to print to console.</param>
         /// <param name="statusCode">Optional HTTP status code to store as a property.</param>
         /// <param name="errorType">Optional error type string to store as a property.</param>
         /// <param name="errors">Optional list of Error objects to store as a property.</param>
-        internal UnexpectedHttpError(string errorMessage, int statusCode, string? errorType = null, List<Error>? errors = null) : base(errorMessage, statusCode, errorType, errors)
+        internal UnknownHttpError(string errorMessage, int statusCode, string? errorType = null, List<Error>? errors = null) : base(errorMessage, statusCode, errorType, errors)
         {
         }
     }