+++ EasyPost/Exceptions/General/FilteringError.cs
@@ -9,5 +9,11 @@
         internal FilteringError(string message) : base(message)
         {
         }
+
+        /// <summary>
+        ///     Get a formatted error string with expanded details about the error.
+        /// </summary>
+        /// <returns>A formatted error string.</returns>
+        public override string PrettyPrint => Message;
     }
 }