+++ EasyPost/Exceptions/General/JsonError.cs
@@ -11,6 +11,12 @@ namespace EasyPost.Exceptions.General
         internal JsonError(string message) : base(message)
         {
         }
+
+        /// <summary>
+        ///     Get a formatted error string with expanded details about the error.
+        /// </summary>
+        /// <returns>A formatted error string.</returns>
+        public override string PrettyPrint => Message;
     }
 
     public class JsonDeserializationError : JsonError