+++ EasyPost/Exceptions/General/SignatureVerificationError.cs
@@ -8,5 +8,11 @@
         internal SignatureVerificationError() : base(Constants.ErrorMessages.InvalidWebhookSignature)
         {
         }
+
+        /// <summary>
+        ///     Get a formatted error string with expanded details about the error.
+        /// </summary>
+        /// <returns>A formatted error string.</returns>
+        public override string PrettyPrint => Message;
     }
 }