+++ EasyPost/Exceptions/_Base.cs
@@ -2,10 +2,16 @@
 
 namespace EasyPost.Exceptions
 {
-    public class EasyPostError : Exception
+    public abstract class EasyPostError : Exception
     {
         internal EasyPostError(string message) : base(message)
         {
         }
+
+        /// <summary>
+        ///     Get a formatted error string with expanded details about the EasyPost API error.
+        /// </summary>
+        /// <returns>A formatted error string.</returns>
+        public abstract string PrettyPrint { get; }
     }
 }